+++ Project/Engine/struct.h
@@ -98,8 +98,8 @@ struct tParticleModule
 	Vec4 vSpawnMinScale; // 초기 최소 크기
 	Vec4 vSpawnMaxScale; // 초기 최대 크기
 
-	int	  SpawnType;		// 0 : Continuous, 1 : Burst
-	int	  BurstInterval;	// Burst 간격(초 단위)
+	int	  SpawnType;	  // 0 : Continuous, 1 : Burst
+	float BurstInterval;  // Burst 간격
 	float MinLife;		  // 최소 수명
 	float MaxLife;		  // 최대 수명
 	float MinMass;		  // 최소 질량
@@ -314,7 +314,7 @@ struct tFontInfo
 
 	// X, Y - 윈도우 좌표 기준 x, y
 	// Z - 깊이 판정을 위한 용도
-	Vec3  vPos;
+	Vec3 vPos;
 
 	float fFontSize; // 폰트의 크기
